+++ src/Service/Auth/AuthCookieStorage.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Ridibooks\Cms\Service\Auth\Authenticator;
+namespace Ridibooks\Cms\Service\Auth;
 
 use Symfony\Component\HttpFoundation\Cookie;
 use Symfony\Component\HttpFoundation\Request;
