+++ src/Service/Auth/OAuth2/Client/AzureClient.php
