+++ src/Service/Auth/OAuth2/Client/OAuth2ClientInterface.php
